fix from file commands printing extra text
